linka backend/services/user-service/app/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
httpx==0.27.2
slowapi==0.1.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic[email]==2.9.2
requests==2.32.3
pytest==7.4.3
pytest-asyncio==0.21.1
supabase==2.3.4
gotrue==2.3.0
